refine api

+++ pkg/service/im/db/db_group_user.go
@@ -8,6 +8,9 @@ import (
 	"context"
 	"fmt"
 
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+
 	"openpitrix.io/iam/pkg/internal/funcutil"
 	"openpitrix.io/iam/pkg/pb/im"
 	"openpitrix.io/iam/pkg/service/im/db_spec"
@@ -17,18 +20,68 @@ import (
 func (p *Database) JoinGroup(ctx context.Context, req *pbim.JoinGroupRequest) (*pbim.Empty, error) {
 	logger.Infof(ctx, funcutil.CallerName(1))
 
+	if len(req.Uid) == 0 || len(req.Gid) == 0 {
+		err := status.Errorf(codes.InvalidArgument, "empty uid or gid")
+		logger.Warnf(ctx, "%+v", err)
+		return nil, err
+	}
+	if !(len(req.Uid) == 1 || len(req.Gid) == 1 || len(req.Uid) == len(req.Gid)) {
+		err := status.Errorf(codes.InvalidArgument,
+			"uid and gid donot math: gid = %v, uid = %v",
+			req.Gid, req.Uid,
+		)
+		logger.Warnf(ctx, "%+v", err)
+		return nil, err
+	}
+
 	sql := fmt.Sprintf(
 		`insert into %s (id, user_id, group_id) values(?,?,?)`,
 		db_spec.UserGroupBindingTableName,
 	)
 
-	xid := genXid()
-	_, err := p.DB.ExecContext(ctx, sql, xid, req.GetUid(), req.GetGid())
-	if err != nil {
-		logger.Warnf(ctx, "%v", sql)
-		logger.Warnf(ctx, "%v, %v, %v", xid, req.GetUid(), req.GetGid())
-		logger.Warnf(ctx, "%+v", err)
-		return nil, err
+	switch {
+	case len(req.Uid) == len(req.Gid):
+		for i := 0; i < len(req.Gid); i++ {
+			xid := genXid()
+			gid := req.Gid[i]
+			uid := req.Uid[i]
+
+			_, err := p.DB.ExecContext(ctx, sql, xid, uid, gid)
+			if err != nil {
+				logger.Warnf(ctx, "%v", sql)
+				logger.Warnf(ctx, "%v, %v, %v", xid, uid, gid)
+				logger.Warnf(ctx, "%+v", err)
+				return nil, err
+			}
+		}
+	case len(req.Uid) == 1:
+		for i := 0; i < len(req.Gid); i++ {
+			xid := genXid()
+			gid := req.Gid[i]
+			uid := req.Uid[0]
+
+			_, err := p.DB.ExecContext(ctx, sql, xid, uid, gid)
+			if err != nil {
+				logger.Warnf(ctx, "%v", sql)
+				logger.Warnf(ctx, "%v, %v, %v", xid, uid, gid)
+				logger.Warnf(ctx, "%+v", err)
+				return nil, err
+			}
+		}
+	case len(req.Gid) == 1:
+		for i := 0; i < len(req.Uid); i++ {
+			xid := genXid()
+			gid := req.Gid[0]
+			uid := req.Uid[i]
+
+			_, err := p.DB.ExecContext(ctx, sql, xid, uid, gid)
+			if err != nil {
+				logger.Warnf(ctx, "%v", sql)
+				logger.Warnf(ctx, "%v, %v, %v", xid, uid, gid)
+				logger.Warnf(ctx, "%+v", err)
+				return nil, err
+			}
+		}
 	}
 
 	return &pbim.Empty{}, nil
@@ -37,16 +90,68 @@ func (p *Database) JoinGroup(ctx context.Context, req *pbim.JoinGroupRequest) (*
 func (p *Database) LeaveGroup(ctx context.Context, req *pbim.LeaveGroupRequest) (*pbim.Empty, error) {
 	logger.Infof(ctx, funcutil.CallerName(1))
 
+	if len(req.Uid) == 0 || len(req.Gid) == 0 {
+		err := status.Errorf(codes.InvalidArgument, "empty uid or gid")
+		logger.Warnf(ctx, "%+v", err)
+		return nil, err
+	}
+	if !(len(req.Uid) == 1 || len(req.Gid) == 1 || len(req.Uid) == len(req.Gid)) {
+		err := status.Errorf(codes.InvalidArgument,
+			"uid and gid donot math: gid = %v, uid = %v",
+			req.Gid, req.Uid,
+		)
+		logger.Warnf(ctx, "%+v", err)
+		return nil, err
+	}
+
 	sql := fmt.Sprintf(
 		`delete from %s where user_id=? AND group_id=?`,
 		db_spec.UserGroupBindingTableName,
 	)
 
-	_, err := p.DB.ExecContext(ctx, sql, req.GetUid(), req.GetGid())
-	if err != nil {
-		logger.Warnf(ctx, "%v", sql)
-		logger.Warnf(ctx, "%+v", err)
-		return nil, err
+	switch {
+	case len(req.Uid) == len(req.Gid):
+		for i := 0; i < len(req.Gid); i++ {
+			xid := genXid()
+			gid := req.Gid[i]
+			uid := req.Uid[i]
+
+			_, err := p.DB.ExecContext(ctx, sql, xid, uid, gid)
+			if err != nil {
+				logger.Warnf(ctx, "%v", sql)
+				logger.Warnf(ctx, "%v, %v, %v", xid, uid, gid)
+				logger.Warnf(ctx, "%+v", err)
+				return nil, err
+			}
+		}
+	case len(req.Uid) == 1:
+		for i := 0; i < len(req.Gid); i++ {
+			xid := genXid()
+			gid := req.Gid[i]
+			uid := req.Uid[0]
+
+			_, err := p.DB.ExecContext(ctx, sql, xid, uid, gid)
+			if err != nil {
+				logger.Warnf(ctx, "%v", sql)
+				logger.Warnf(ctx, "%v, %v, %v", xid, uid, gid)
+				logger.Warnf(ctx, "%+v", err)
+				return nil, err
+			}
+		}
+	case len(req.Gid) == 1:
+		for i := 0; i < len(req.Uid); i++ {
+			xid := genXid()
+			gid := req.Gid[0]
+			uid := req.Uid[i]
+
+			_, err := p.DB.ExecContext(ctx, sql, xid, uid, gid)
+			if err != nil {
+				logger.Warnf(ctx, "%v", sql)
+				logger.Warnf(ctx, "%v, %v, %v", xid, uid, gid)
+				logger.Warnf(ctx, "%+v", err)
+				return nil, err
+			}
+		}
 	}
 
 	return &pbim.Empty{}, nil
